coal_detection_app/requirements.txt
Flask
torch
torchvision
Pillow
numpy
opencv-python
Werkzeug
gunicorn